fix: change SecurityVersion type to uint16 for SGX enclave compatibility (ref #1)

+++ info.go
@@ -22,7 +22,7 @@ type SgxAleoInfo struct {
 
 // Contains information about an SGX enclave.
 type SgxInfo struct {
-	SecurityVersion uint        `json:"securityVersion"` // Security version of the enclave. For SGX enclaves, this is the ISVSVN value.
+	SecurityVersion uint16        `json:"securityVersion"` // Security version of the enclave. For SGX enclaves, this is the ISVSVN value.
 	Debug           bool        `json:"debug"`           // If true, the report is for a debug enclave.
 	UniqueID        string      `json:"uniqueId"`        // The unique ID for the enclave. For SGX enclaves, this is the MRENCLAVE value.
 	SignerID        string      `json:"signerId"`        // The signer ID for the enclave. For SGX enclaves, this is the MRSIGNER value.
